+++ hooks/index.ts
@@ -3,6 +3,8 @@ export { useGoals } from './useGoals'
 export { useGoalsCopy } from './useGoalsCopy'
 export { useDaily } from './useDaily'
 export { useForecast } from './useForecast'
+export { useInlineEdit } from './useInlineEdit'
+export { useCopyDropdown } from './useCopyDropdown'
 
 // Re-export types
 export type { DreamGoal, Goal, GoalTag } from './useGoals'